+++ babel.config.js
@@ -1,14 +1,3 @@
 module.exports = {
-  presets: [
-    [
-      '@babel/preset-env',
-      {
-        targets: {
-          node: '14',
-        },
-        modules: 'auto',
-      },
-    ],
-  ],
-  plugins: ['@babel/plugin-proposal-export-namespace-from'],
+  presets: [['@babel/preset-env', { targets: { node: 'current' } }]],
 };