+++ src/config.js
@@ -21,8 +21,6 @@ export default {
   stage: process.env.STAGE,
   // The log level
   logLevel: process.env.LOG_LEVEL,
-  // The auth secret
-  swsSecret: process.env.SWS_SECRET,
   // The storage bucket
   audioBucketName: process.env.AUDIO_BUCKET_NAME,
   // The CORS allowed origins
@@ -37,10 +35,6 @@ export default {
 
   m3u8FileTableTTL: parseInt(process.env.AUDIO_CACHE_TTL, 10) || 60 * 60 * 24 * 30,
 
-  cloudfrontDomain: process.env.AWS_CLOUDFRONT_DOMAIN,
-  cloudfrontKeypairId: process.env.AWS_CLOUDFRONT_KEYPAIR_ID,
-  cloudfrontPrivateKeyString: process.env.AWS_CLOUDFRONT_PRIVATE_KEY?.replace(/\\n/g, '\n'),
-
   signedUrlExpiresIn: parseInt(process.env.SIGNED_URL_EXPIRY_SEC, 10) || 3600,
 
   createM3U8LambdaArn: process.env.LAMBDA_CREATE_M3U8_ARN,
@@ -52,4 +46,7 @@ export default {
   validAudioFormats: parseCsv(process.env.ALLOWED_FORMATS) || ['mp3'],
   validSegmentTimes: parseCsv(process.env.ALLOWED_SEGMENT_DURATION) || [5],
   allowedAudioOrigins: parseCsv(process.env.ALLOWED_AUDIO_ORIGINS) || ['*'],
+
+  // this is to limit output in case of excessively long tracks
+  audioOutputMaxDuration: parseInt(process.env.AUDIO_OUTPUT_MAX_DURATION, 10) || 600, // 10 mins
 };