+++ src/libs/getTmpFolder.js
@@ -15,7 +15,7 @@
  * along with this program. If not, see <https://www.gnu.org/licenses/>.
  */
 import { resolve } from 'path';
-import { v4 as createUuid } from 'uuid';
+import { randomUUID } from 'crypto';
 import { mkdir } from 'fs/promises';
 import { existsSync } from 'fs';
 
@@ -26,7 +26,7 @@ import { existsSync } from 'fs';
  * @param {String} params.basedir - the basedir to use to create the tmp folder
  */
 export default async ({ basedir = '/tmp' } = {}) => {
-  const tmpWorkspaceFolder = resolve(basedir, createUuid());
+  const tmpWorkspaceFolder = resolve(basedir, randomUUID());
   if (!existsSync(tmpWorkspaceFolder)) await mkdir(tmpWorkspaceFolder, { recursive: true });
   return tmpWorkspaceFolder;
 };