+++ src/services/audioSegment.js
@@ -16,10 +16,11 @@
  */
 import { readdir, readFile, writeFile, unlink } from 'fs/promises';
 import { resolve } from 'path';
-import { logger } from '@soundws/service-libs';
+import logger from '@soundws/service-libs/src/logger';
 import convertWavToFormat from '../libs/audioConvertWavToFormat';
 import audioSegmentToWav from '../libs/audioSegmentToWav';
 import getTmpFolder from '../libs/getTmpFolder';
+import config from '../config';
 
 export default async (options) => {
   try {
@@ -31,7 +32,11 @@ export default async (options) => {
     });
 
     // 1 first we simply segment the wavs without converting to the target format
-    await audioSegmentToWav({ tmpWorkspaceFolder, ...options });
+    await audioSegmentToWav({
+      tmpWorkspaceFolder,
+      audioOutputMaxDuration: config.audioOutputMaxDuration, // ensure we dont produce overly large files
+      ...options,
+    });
 
     // Get the wavs that have been generated
     const files = (await readdir(tmpWorkspaceFolder))