+++ src/services/lambdaInvokeCreateM3u8.js
@@ -14,24 +14,17 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program. If not, see <https://www.gnu.org/licenses/>.
  */
-import { createHandler, createPolicy, verifyAccessToken, logger } from '@soundws/service-libs';
+import lambdaInvoke from '@soundws/service-libs/src/lambdaInvoke';
 import config from '../config';
 
-const handler = createHandler(async (event) => {
-  logger.debug('Running with config', config);
-
-  try {
-    const token = event?.authorizationToken?.replace('Bearer ', '');
-    if (!token) throw new Error('no bearer token in request');
-
-    const decoded = verifyAccessToken(token, config.swsSecret, {
-      audience: config.jwtAudience,
-    });
-    return createPolicy(event, 'Allow', decoded.sub, decoded);
-  } catch (error) {
-    return createPolicy(event, 'Deny');
-  }
-});
-
-// eslint-disable-next-line import/prefer-default-export
-export { handler };
+/**
+ * @param {Object} event - the event to pass to the lambda
+ */
+export default (event) =>
+  lambdaInvoke({
+    FunctionName: config.createM3U8LambdaArn,
+    InvocationType: 'RequestResponse',
+    Payload: JSON.stringify({
+      queryStringParameters: event.queryStringParameters,
+    }),
+  });