+++ tests/libs/getTmpFolder.test.js
@@ -6,7 +6,7 @@ describe('getTmpFolder', () => {
     const tmpFolder = await getTmpFolder();
 
     expect(tmpFolder).toMatch(
-      /\/tmp\/[0-9a-fA-F]{8}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{12}/
+      /\/tmp\/[0-9a-fA-F]{8}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{12}/,
     );
 
     rmdirSync(tmpFolder);
@@ -16,7 +16,7 @@ describe('getTmpFolder', () => {
     const tmpFolder = await getTmpFolder({ basedir: '/tmp/somewhereelse' });
 
     expect(tmpFolder).toMatch(
-      /\/tmp\/somewhereelse\/[0-9a-fA-F]{8}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{12}/
+      /\/tmp\/somewhereelse\/[0-9a-fA-F]{8}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{12}/,
     );
 
     rmdirSync(tmpFolder);